[BinFile] Update Segment type
To use less memory

+++ src/FrontEnd/BinFile/PEHelper.fs
@@ -239,9 +239,9 @@ let getSecPermission (chr: SectionCharacteristics) =
 let getSegments pe =
   let secToSegment (sec: SectionHeader) =
     { Address = uint64 sec.VirtualAddress + pe.BaseAddr
-      Offset = uint64 sec.PointerToRawData
-      Size = getVirtualSectionSize sec |> uint64
-      SizeInFile = uint64 sec.SizeOfRawData
+      Offset = uint32 sec.PointerToRawData
+      Size = getVirtualSectionSize sec |> uint32
+      SizeInFile = uint32 sec.SizeOfRawData
       Permission = getSecPermission sec.SectionCharacteristics }
   pe.SectionHeaders
   |> Seq.map secToSegment
